Extract xaml header and closer generation to a separate class, that is going to be reused by Simple and Rich layout generators
